pre merge

+++ Tree.cpp
@@ -198,7 +198,7 @@ int Tree::CalcFormula(Node* n, bool x[], std::vector<std::string> xlabels){
   }
 }
 
-int Tree::CalcFitness(Node* n, bool x[], std::vector<std::string> xlabels, int y){
+float Tree::CalcFitness(Node* n, bool x[], std::vector<std::string> xlabels, int y){
   return CalcFormula(n,x,xlabels) - y;
 }
